split people code
