+++ webui-src/app/people/people_state.js
@@ -1,0 +1,376 @@
+const m = require('mithril');
+const rs = require('rswebui');
+const Data = require('network/network_data');
+const peopleUtil = require('people/people_util');
+
+const State = {
+  searchString: '',
+  selectedId: null, // GXS ID of the selected identity
+  activeFilter: 'contacts', // 'all' | 'contacts' | 'own'
+  gxsIdToDetailsMap: {},
+  ownGxsIds: [],
+  gpgToGxsIdMap: {},
+  showMailCompose: false,
+  activeTab: 'details',
+  selectedOwnGxsIdForChat: '',
+  chatPid: null,
+  chatMessages: [],
+  chatInputMsg: '',
+  distantChatStatus: null,
+  statusPollInterval: null,
+  chatDisconnected: false,
+  activeMenu: null,
+};
+
+function fetchIdDetails(gxsId) {
+  if (!gxsId) return;
+  if (State.gxsIdToDetailsMap[gxsId] === undefined) {
+    State.gxsIdToDetailsMap[gxsId] = null; // Mark as loading
+    rs.rsJsonApiRequest('/rsIdentity/getIdDetails', { id: gxsId }, (detData) => {
+      if (detData && detData.details) {
+        State.gxsIdToDetailsMap[gxsId] = detData.details;
+        const pgpId = detData.details.mPgpId;
+        if (pgpId && pgpId !== '0000000000000000') {
+          State.gpgToGxsIdMap[pgpId.toLowerCase()] = gxsId;
+        }
+        m.redraw();
+      }
+    });
+  }
+}
+
+function loadGxsIdentities() {
+  rs.rsJsonApiRequest('/rsIdentity/getIdentitiesSummaries', {}, (data) => {
+    if (data && data.ids) {
+      m.redraw();
+    }
+  });
+}
+
+function loadOwnGxsIds() {
+  return new Promise((resolve) => {
+    peopleUtil.ownIds((ids) => {
+      State.ownGxsIds = ids || [];
+      if (State.ownGxsIds.length > 0 && !State.selectedOwnGxsIdForChat) {
+        State.selectedOwnGxsIdForChat = State.ownGxsIds[0];
+      }
+      m.redraw();
+      resolve();
+    });
+  });
+}
+
+function get64Num(val) {
+  if (!val) return 0;
+  if (typeof val === 'object') {
+    return val.xint64 || parseInt(val.xstr64) || 0;
+  }
+  return Number(val) || 0;
+}
+
+function getServiceName(serviceId) {
+  switch (serviceId) {
+    case 1: return 'Channels';
+    case 2: return 'Forums';
+    case 3: return 'Boards';
+    case 4: return 'Chat';
+    case 5: return 'GxsCircles';
+    case 6: return 'GxsMail';
+    case 7: return 'GxsCircles';
+    case 8: return 'Wire';
+    default: return 'Unknown (' + serviceId + ')';
+  }
+}
+
+function createUsageString(u) {
+  if (!u) return '[Unknown]';
+  const serviceName = getServiceName(u.mServiceId);
+  const usageCode = u.mUsageCode;
+
+  switch (usageCode) {
+    case 0:
+      return '[Unknown]';
+    case 1:
+      return `Admin signature in service ${serviceName}`;
+    case 2:
+      return `Admin signature verification in service ${serviceName}`;
+    case 3:
+      return `Creation of author signature in service ${serviceName}`;
+    case 4:
+    case 7:
+      return `Group author for group ${u.mGrpId || 'Unknown'} in service ${serviceName}`;
+    case 5:
+      return `Message signature creation in group ${u.mGrpId || 'Unknown'} of service ${serviceName}`;
+    case 6:
+    case 8:
+      return `Vote/comment in ${serviceName} service (Group: ${u.mGrpId || 'Unknown'}, Msg: ${u.mMsgId || 'Unknown'})`;
+    case 9:
+      return `Message in chat room (Id: ${get64Num(u.mAdditionalId)})`;
+    case 10:
+      return 'Distant message signature validation.';
+    case 11:
+      return 'Distant message signature creation.';
+    case 12:
+      return 'Signature validation in distant tunnel system.';
+    case 13:
+      return 'Signature in distant tunnel system.';
+    case 14:
+      return 'Received from GXS sync.';
+    case 15:
+      return 'Received from GXS discovery.';
+    case 16:
+      return 'Explicit request to friend.';
+    case 17:
+      return 'Generic signature validation.';
+    case 18:
+      return 'Generic signature creation.';
+    case 19:
+      return 'Generic encryption.';
+    case 20:
+      return 'Generic decryption.';
+    case 21:
+      return 'Circle membership check.';
+    default:
+      return `Usage code ${usageCode} in service ${serviceName}`;
+  }
+}
+
+function getSafeAvatar(details) {
+  return details && details.mAvatar ? details.mAvatar : undefined;
+}
+
+function getOnlineSslId(gpgId) {
+  if (!gpgId) return null;
+  const friend = Data.gpgDetails[gpgId.toLowerCase()];
+  if (friend && friend.locations) {
+    const onlineLoc = friend.locations.find((loc) => loc.isOnline);
+    return onlineLoc ? onlineLoc.id : null;
+  }
+  return null;
+}
+
+function isIdentityOnline(gxsId) {
+  fetchIdDetails(gxsId);
+  const details = State.gxsIdToDetailsMap[gxsId];
+  if (details && details.mPgpId && details.mPgpId !== '0000000000000000') {
+    const friend = Data.gpgDetails[details.mPgpId.toLowerCase()];
+    return friend ? friend.isOnline : false;
+  }
+  return false;
+}
+
+function syncFilter(tab) {
+  let newFilter = 'all';
+  if (tab === 'OwnIdentity') {
+    newFilter = 'own';
+  } else if (tab === 'MyContacts') {
+    newFilter = 'contacts';
+  }
+
+  if (State.activeFilter !== newFilter) {
+    State.activeFilter = newFilter;
+    State.selectedId = null;
+    State.chatPid = null;
+    State.chatMessages = [];
+    State.chatInputMsg = '';
+    State.activeTab = 'details';
+  }
+}
+
+function getStatusColor(status) {
+  switch (status) {
+    case 1: return '#eab308'; // Yellow
+    case 2: return '#22c55e'; // Green
+    case 3: return '#ef4444'; // Red
+    default: return '#94a3b8'; // Grey
+  }
+}
+
+function getStatusTooltip(status) {
+  switch (status) {
+    case 1: return 'Tunnel is pending. Please wait...';
+    case 2: return 'End-to-end encrypted conversation established. You can talk!';
+    case 3: return 'Your partner closed the conversation.';
+    default: return 'Remote status unknown.';
+  }
+}
+
+function pollDistantChatStatus() {
+  if (!State.chatPid) return;
+  rs.rsJsonApiRequest(
+    '/rsChats/getDistantChatStatus',
+    {
+      pid: State.chatPid,
+    },
+    (detail, success) => {
+      if (success && detail.retval) {
+        const oldStatus = State.distantChatStatus ? State.distantChatStatus.status : null;
+        State.distantChatStatus = detail.info;
+
+        if (oldStatus !== null && oldStatus !== detail.info.status) {
+          if (detail.info.status === 2) {
+            const text = 'Tunnel is secured. You can talk!';
+            const exists = State.chatMessages.some((m) => m.isSystem && m.msg === text);
+            if (!exists) {
+              State.chatMessages.push({
+                incoming: true,
+                isSystem: true,
+                msg: text,
+                sendTime: Math.floor(Date.now() / 1000),
+              });
+              State.chatMessages.sort((a, b) => a.sendTime - b.sendTime);
+            }
+          } else if (detail.info.status === 3) {
+            const text = 'Your partner closed the conversation.';
+            const exists = State.chatMessages.some((m) => m.isSystem && m.msg === text);
+            if (!exists) {
+              State.chatMessages.push({
+                incoming: true,
+                isSystem: true,
+                msg: text,
+                sendTime: Math.floor(Date.now() / 1000),
+              });
+              State.chatMessages.sort((a, b) => a.sendTime - b.sendTime);
+            }
+          }
+        }
+        m.redraw();
+      }
+    }
+  );
+}
+
+function startStatusPolling() {
+  stopStatusPolling();
+  pollDistantChatStatus();
+  State.statusPollInterval = setInterval(pollDistantChatStatus, 3000);
+}
+
+function stopStatusPolling() {
+  if (State.statusPollInterval) {
+    clearInterval(State.statusPollInterval);
+    State.statusPollInterval = null;
+  }
+  State.distantChatStatus = null;
+}
+
+function initializeDistantChat() {
+  if (!State.selectedId || !State.selectedOwnGxsIdForChat) return;
+
+  State.chatPid = null;
+  State.chatMessages = [];
+  State.chatDisconnected = false;
+  m.redraw();
+
+  rs.rsJsonApiRequest(
+    '/rsChats/initiateDistantChatConnexion',
+    {
+      to_pid: State.selectedId,
+      from_pid: State.selectedOwnGxsIdForChat,
+      notify: true,
+    },
+    (res) => {
+      if (res && res.pid) {
+        State.chatPid = rs.idToHex(res.pid);
+        State.distantChatStatus = null;
+        loadChatMessages();
+        pollDistantChatStatus();
+        startStatusPolling();
+      }
+    }
+  );
+}
+
+function loadChatMessages() {
+  if (!State.chatPid) return;
+
+  const chatPeerId = {
+    broadcast_status_peer_id: '00000000000000000000000000000000',
+    type: 2, // DISTANT
+    peer_id: '00000000000000000000000000000000',
+    distant_chat_id: State.chatPid,
+    lobby_id: { xstr64: '0' },
+  };
+
+  rs.rsJsonApiRequest(
+    '/rsHistory/getMessages',
+    {
+      chatPeerId: chatPeerId,
+      loadCount: 50,
+    },
+    (data, success) => {
+      if (success && data.msgs) {
+        State.chatMessages = data.msgs;
+        m.redraw();
+        setTimeout(() => {
+          const element = document.querySelector('.chat-messages');
+          if (element) element.scrollTop = element.scrollHeight;
+        }, 100);
+      }
+    }
+  );
+}
+
+function sendDistantChatMessage() {
+  if (!State.chatInputMsg.trim() || !State.chatPid) return;
+
+  const cid = {
+    broadcast_status_peer_id: '00000000000000000000000000000000',
+    type: 2, // DISTANT
+    peer_id: '00000000000000000000000000000000',
+    distant_chat_id: State.chatPid,
+    lobby_id: { xstr64: '0' },
+  };
+
+  const text = State.chatInputMsg;
+  State.chatInputMsg = '';
+
+  const echoMsg = {
+    chat_id: cid,
+    msg: text,
+    sendTime: Math.floor(Date.now() / 1000),
+    incoming: false,
+    lobby_peer_gxs_id: State.selectedOwnGxsIdForChat,
+  };
+  State.chatMessages.push(echoMsg);
+  m.redraw();
+  setTimeout(() => {
+    const element = document.querySelector('.chat-messages');
+    if (element) element.scrollTop = element.scrollHeight;
+  }, 100);
+
+  rs.rsJsonApiRequest(
+    '/rsChats/sendChat',
+    {
+      id: cid,
+      msg: text,
+    },
+    (data, success) => {
+      if (!success) {
+        console.error('[RS] Failed to send distant chat message');
+      }
+    }
+  );
+}
+
+module.exports = {
+  State,
+  fetchIdDetails,
+  loadGxsIdentities,
+  loadOwnGxsIds,
+  get64Num,
+  getServiceName,
+  createUsageString,
+  getSafeAvatar,
+  getOnlineSslId,
+  isIdentityOnline,
+  syncFilter,
+  getStatusColor,
+  getStatusTooltip,
+  pollDistantChatStatus,
+  startStatusPolling,
+  stopStatusPolling,
+  initializeDistantChat,
+  loadChatMessages,
+  sendDistantChatMessage,
+};
